Minor test rework after improving startup, a lot of mocks weren't being used correctly.
